devfolio-backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
django-filter==23.5
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
